+++ src/Streaming/Prelude.hs
@@ -47,9 +47,17 @@
 > --------------------------------------------------------------------------------------------------------------------
 >
 -}
-{-# LANGUAGE RankNTypes, BangPatterns, DeriveDataTypeable, TypeFamilies,
-             DeriveFoldable, DeriveFunctor, DeriveTraversable, CPP,
-             ScopedTypeVariables, Trustworthy #-}
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveFoldable #-}
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE DeriveTraversable #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies #-}
+
+{-# OPTIONS_GHC -Wall #-}
 
 module Streaming.Prelude (
     -- * Types
@@ -106,6 +114,7 @@ module Streaming.Prelude (
     , sequence
     , filter
     , filterM
+    , mapMaybeM
     , delay
     , intersperse
     , take
@@ -231,12 +240,10 @@ module Streaming.Prelude (
 
     -- * Basic Type
     , Stream
-
   ) where
 import Streaming.Internal
 
 import Control.Monad hiding (filterM, mapM, mapM_, foldM, foldM_, replicateM, sequence)
-import Data.Data ( Data, Typeable )
 import Data.Functor.Identity
 import Data.Functor.Sum
 import Control.Monad.Trans
@@ -244,8 +251,6 @@ import Control.Applicative (Applicative (..))
 import Data.Functor (Functor (..), (<$))
 
 import qualified Prelude as Prelude
-import Data.Foldable (Foldable)
-import Data.Traversable (Traversable)
 import qualified Data.Foldable as Foldable
 import qualified Data.Sequence as Seq
 import Text.Read (readMaybe)
@@ -254,7 +259,7 @@ import Prelude hiding (map, mapM, mapM_, filter, drop, dropWhile, take, mconcat
                       , takeWhile, enumFrom, enumFromTo, enumFromThen, length
                       , print, zipWith, zip, zipWith3, zip3, unzip, seq, show, read
                       , readLn, sequence, concat, span, break, readFile, writeFile
-                      , minimum, maximum, elem, notElem, intersperse, all, any, head
+                      , minimum, maximum, elem, notElem, all, any, head
                       , last)
 
 import qualified GHC.IO.Exception as G
@@ -262,12 +267,8 @@ import qualified System.IO as IO
 import Foreign.C.Error (Errno(Errno), ePIPE)
 import Control.Exception (throwIO, try)
 import Data.Monoid (Monoid (mappend, mempty))
-import Data.String (IsString (..))
 import Control.Concurrent (threadDelay)
-import Data.Functor.Classes
 import Data.Functor.Compose
-
-import GHC.Magic
 import Data.Functor.Of
 
 -- instance (Eq a) => Eq1 (Of a) where eq1 = (==)
@@ -335,10 +336,10 @@ strictly = \(a,b) -> a :> b
  -}
 
 fst' :: Of a b -> a
-fst' (a :> b) = a
+fst' (a :> _) = a
 {-#INLINE fst' #-}
 snd' :: Of a b -> b
-snd' (a :> b) = b
+snd' (_ :> b) = b
 {-#INLINE snd' #-}
 
 {-| Map a function over the first element of an @Of@ pair
@@ -388,7 +389,7 @@ all thus = loop True where
 all_ :: Monad m => (a -> Bool) -> Stream (Of a) m r -> m Bool
 all_ thus = loop True where
   loop b str = case str of
-    Return r -> return b
+    Return _ -> return b
     Effect m -> m >>= loop b
     Step (a :> rest) -> if thus a
       then loop True rest
@@ -411,7 +412,7 @@ any thus = loop False where
 any_ :: Monad m => (a -> Bool) -> Stream (Of a) m r -> m Bool
 any_ thus = loop False where
   loop b str = case str of
-    Return r -> return b
+    Return _ -> return b
     Effect m -> m >>= loop b
     Step (a :> rest) -> if thus a
       then return True
@@ -432,11 +433,11 @@ any_ thus = loop False where
 
 break :: Monad m => (a -> Bool) -> Stream (Of a) m r
       -> Stream (Of a) m (Stream (Of a) m r)
-break pred = loop where
+break thePred = loop where
   loop str = case str of
     Return r         -> Return (Return r)
-    Effect m          -> Effect $ liftM loop m
-    Step (a :> rest) -> if (pred a)
+    Effect m          -> Effect $ fmap loop m
+    Step (a :> rest) -> if (thePred a)
       then Return (Step (a :> rest))
       else Step (a :> loop rest)
 {-# INLINABLE break #-}
@@ -461,18 +462,18 @@ break pred = loop where
 
 -}
 breakWhen :: Monad m => (x -> a -> x) -> x -> (x -> b) -> (b -> Bool) -> Stream (Of a) m r -> Stream (Of a) m (Stream (Of a) m r)
-breakWhen step begin done pred = loop0 begin
+breakWhen step begin done thePred = loop0 begin
   where
     loop0 x stream = case stream of
         Return r -> return (return r)
-        Effect mn  -> Effect $ liftM (loop0 x) mn
+        Effect mn  -> Effect $ fmap (loop0 x) mn
         Step (a :> rest) -> loop a (step x a) rest
     loop a !x stream = do
-      if pred (done x)
+      if thePred (done x)
         then return (yield a >> stream)
         else case stream of
           Return r -> yield a >> return (return r)
-          Effect mn  -> Effect $ liftM (loop a x) mn
+          Effect mn  -> Effect $ fmap (loop a x) mn
           Step (a' :> rest) -> do
             yield a
             loop a' (step x a') rest
@@ -518,7 +519,7 @@ chain :: Monad m => (a -> m ()) -> Stream (Of a) m r -> Stream (Of a) m r
 chain f = loop where
   loop str = case str of
     Return r -> return r
-    Effect mn  -> Effect (liftM loop mn)
+    Effect mn  -> Effect (fmap loop mn)
     Step (a :> rest) -> Effect $ do
       f a
       return (Step (a :> loop rest))
@@ -661,10 +662,10 @@ drop :: (Monad m) => Int -> Stream (Of a) m r -> Stream (Of a) m r
 drop n str | n <= 0 = str
 drop n str = loop n str where
   loop 0 stream = stream
-  loop n stream = case stream of
+  loop m stream = case stream of
       Return r       -> Return r
-      Effect ma      -> Effect (liftM (loop n) ma)
-      Step (a :> as) -> loop (n-1) as
+      Effect ma      -> Effect (fmap (loop m) ma)
+      Step (_ :> as) -> loop (m-1) as
 {-# INLINABLE drop #-}
 
 -- ---------------
@@ -685,11 +686,11 @@ four<Enter>
 
 -}
 dropWhile :: Monad m => (a -> Bool) -> Stream (Of a) m r -> Stream (Of a) m r
-dropWhile pred = loop where
+dropWhile thePred = loop where
   loop stream = case stream of
     Return r       -> Return r
-    Effect ma       -> Effect (liftM loop ma)
-    Step (a :> as) -> if pred a
+    Effect ma       -> Effect (fmap loop ma)
+    Step (a :> as) -> if thePred a
       then loop as
       else Step (a :> as)
 {-# INLINABLE dropWhile #-}
@@ -746,21 +747,21 @@ effects = loop where
 
 elem :: (Monad m, Eq a) => a -> Stream (Of a) m r -> m (Of Bool r)
 elem a' = loop False where
-  loop True str = liftM (True :>) (effects str)
+  loop True str = fmap (True :>) (effects str)
   loop False str = case str of
     Return r -> return (False :> r)
     Effect m -> m >>= loop False
     Step (a:> rest) ->
       if a == a'
-        then liftM (True :>) (effects rest)
+        then fmap (True :>) (effects rest)
         else loop False rest
 {-#INLINABLE elem #-}
 
 elem_ :: (Monad m, Eq a) => a -> Stream (Of a) m r -> m Bool
 elem_ a' = loop False where
-  loop True str = return True
+  loop True _ = return True
   loop False str = case str of
-    Return r -> return False
+    Return _ -> return False
     Effect m -> m >>= loop False
     Step (a:> rest) ->
       if a == a'
@@ -822,8 +823,8 @@ erase :: Monad m => Stream (Of a) m r -> Stream Identity m r
 erase = loop where
   loop str = case str of
     Return r -> Return r
-    Effect m -> Effect (liftM loop m)
-    Step (a:>rest) -> Step (Identity (loop rest))
+    Effect m -> Effect (fmap loop m)
+    Step (_:>rest) -> Step (Identity (loop rest))
 {-# INLINABLE erase #-}
 
 -- ---------------
@@ -832,13 +833,13 @@ erase = loop where
 
 -- | Skip elements of a stream that fail a predicate
 filter  :: (Monad m) => (a -> Bool) -> Stream (Of a) m r -> Stream (Of a) m r
-filter pred = loop where
+filter thePred = loop where
   loop str = case str of
     Return r       -> Return r
-    Effect m       -> Effect (liftM loop m)
-    Step (a :> as) -> if pred a
-                         then Step (a :> loop as)
-                         else loop as
+    Effect m       -> Effect (fmap loop m)
+    Step (a :> as) -> if thePred a
+      then Step (a :> loop as)
+      else loop as
 {-# INLINE filter #-}  -- ~ 10% faster than INLINABLE in simple bench
 
                          
@@ -848,12 +849,12 @@ filter pred = loop where
 
 -- | Skip elements of a stream that fail a monadic test
 filterM  :: (Monad m) => (a -> m Bool) -> Stream (Of a) m r -> Stream (Of a) m r
-filterM pred = loop where
+filterM thePred = loop where
   loop str = case str of
     Return r       -> Return r
-    Effect m       -> Effect $ liftM loop m
+    Effect m       -> Effect $ fmap loop m
     Step (a :> as) -> Effect $ do
-      bool <- pred a
+      bool <- thePred a
       if bool
         then return $ Step (a :> loop as)
         else return $ loop as
@@ -932,7 +933,7 @@ filterM pred = loop where
 > Control.Foldl.purely fold :: Monad m => Fold a b -> Stream (Of a) m () -> m b
 -}
 fold_ :: Monad m => (x -> a -> x) -> x -> (x -> b) -> Stream (Of a) m r -> m b
-fold_ step begin done = liftM (\(a:>rest) -> a) . fold step begin done
+fold_ step begin done = fmap (\(a :> _) -> a) . fold step begin done
 {-#INLINE fold_ #-}
 
 {-| Strict fold of a 'Stream' of elements that preserves the return value.
@@ -984,7 +985,7 @@ fold step begin done str =  fold_loop str begin
 foldM_
     :: Monad m
     => (x -> a -> m x) -> m x -> (x -> m b) -> Stream (Of a) m r -> m b
-foldM_ step begin done  = liftM (\(a :> rest) -> a) . foldM step begin done
+foldM_ step begin done = fmap (\(a :> _) -> a) . foldM step begin done
 {-#INLINE foldM_ #-}
 
 {-| Strict, monadic fold of the elements of a 'Stream (Of a)'
@@ -1070,10 +1071,8 @@ for :: (Monad m, Functor f) => Stream (Of a) m r -> (a -> Stream f m x) -> Strea
 for str0 act = loop str0 where
   loop str = case str of
     Return r         -> Return r
-    Effect m          -> Effect $ liftM loop m
-    Step (a :> rest) -> do
-      act a
-      loop rest
+    Effect m          -> Effect $ fmap loop m
+    Step (a :> rest) -> act a *> loop rest
 {-# INLINABLE for #-}
 
 -- -| Group layers of any functor by comparisons on a preliminary annotation
@@ -1092,11 +1091,11 @@ for str0 act = loop str0 where
 --                 fmap loop (Step $ Compose (a :> fmap (span' (equals a)) p'))
 --   span' :: (Monad m, Functor f) => (a -> Bool) -> Stream (Compose (Of a) f) m r
 --         -> Stream (Compose (Of a) f) m (Stream (Compose (Of a) f) m r)
---   span' pred = loop where
+--   span' thePred = loop where
 --     loop str = case str of
 --       Return r         -> Return (Return r)
---       Effect m          -> Effect $ liftM loop m
---       Step s@(Compose (a :> rest)) -> case pred a  of
+--       Effect m          -> Effect $ fmap loop m
+--       Step s@(Compose (a :> rest)) -> case thePred a  of
 --         True  -> Step (Compose (a :> fmap loop rest))
 --         False -> Return (Step s)
 -- {-# INLINABLE groupedBy #-}
@@ -1151,20 +1150,20 @@ head str = case str of
 
 head_ :: Monad m => Stream (Of a) m r -> m (Maybe a)
 head_ str = case str of
-  Return r            -> return Nothing
-  Effect m            -> m >>= head_
-  Step (a :> rest)    -> return (Just a)
+  Return _ -> return Nothing
+  Effect m -> m >>= head_
+  Step (a :> _) -> return (Just a)
 {-#INLINABLE head_ #-}
 
 intersperse :: Monad m => a -> Stream (Of a) m r -> Stream (Of a) m r
 intersperse x str = case str of
     Return r -> Return r
-    Effect m -> Effect (liftM (intersperse x) m)
+    Effect m -> Effect (fmap (intersperse x) m)
     Step (a :> rest) -> loop a rest
   where
-  loop !a str = case str of
+  loop !a theStr = case theStr of
     Return r -> Step (a :> Return r)
-    Effect m -> Effect (liftM (loop a) m)
+    Effect m -> Effect (fmap (loop a) m)
     Step (b :> rest) -> Step (a :> Step (x :> loop b rest))
 {-#INLINABLE intersperse #-}
 
@@ -1210,11 +1209,11 @@ last = loop Nothing_ where
 last_ :: Monad m => Stream (Of a) m r -> m (Maybe a)
 last_ = loop Nothing_ where
   loop mb str = case str of
-    Return r            -> case mb of
+    Return _ -> case mb of
       Nothing_ -> return Nothing
       Just_ a  -> return (Just a)
-    Effect m            -> m >>= loop mb
-    Step (a :> rest)  -> loop (Just_ a) rest
+    Effect m -> m >>= loop mb
+    Step (a :> rest) -> loop (Just_ a) rest
 {-#INLINABLE last_ #-}
 
 
@@ -1261,7 +1260,7 @@ map f =  maps (\(x :> rest) -> f x :> rest)
 -- loop where  --
   -- loop stream = case stream of
   --   Return r -> Return r
-  --   Effect m -> Effect (liftM loop m)
+  --   Effect m -> Effect (fmap loop m)
   --   Step (a :> as) -> Step (f a :> loop as)
 {-# INLINABLE map #-}
 -- {-# NOINLINE [1] map #-}
@@ -1284,7 +1283,7 @@ mapM :: Monad m => (a -> m b) -> Stream (Of a) m r -> Stream (Of b) m r
 mapM f = loop where
   loop str = case str of
     Return r       -> Return r
-    Effect m        -> Effect (liftM loop m)
+    Effect m        -> Effect (fmap loop m)
     Step (a :> as) -> Effect $ do
       a' <- f a
       return (Step (a' :> loop as) )
@@ -1311,11 +1310,9 @@ mapM f = loop where
 mapM_ :: Monad m => (a -> m b) -> Stream (Of a) m r -> m r
 mapM_ f = loop where
   loop str = case str of
-    Return r       -> return r
-    Effect m        -> m >>= loop
-    Step (a :> as) -> do
-      f a
-      loop as
+    Return r -> return r
+    Effect m -> m >>= loop
+    Step (a :> as) -> f a *> loop as
 {-# INLINABLE mapM_ #-}
 
 
@@ -1334,8 +1331,8 @@ mapM_ f = loop where
 > mapped return        = id
 > maps f . maps g      = maps (f . g)
 > mapped f . mapped g  = mapped (f <=< g)
-> maps f . mapped g    = mapped (liftM f . g)
-> mapped f . maps g    = mapped (f <=< liftM g)
+> maps f . mapped g    = mapped (fmap f . g)
+> mapped f . maps g    = mapped (f <=< fmap g)
 
      @maps@ is more fundamental than @mapped@, which is best understood as a convenience
      for effecting this frequent composition:
@@ -1414,8 +1411,8 @@ maximum_ = fold_ (\m a -> case m of Nothing_ -> Just_ a ; Just_ a' -> Just_ (max
 
      Similarly:
 
-> IOStreams.unfoldM (liftM (either (const Nothing) Just) . next) :: Stream (Of a) IO b -> IO (InputStream a)
-> Conduit.unfoldM (liftM (either (const Nothing) Just) . next)   :: Stream (Of a) m r -> Source a m r
+> IOStreams.unfoldM (fmap (either (const Nothing) Just) . next) :: Stream (Of a) IO b -> IO (InputStream a)
+> Conduit.unfoldM (fmap (either (const Nothing) Just) . next)   :: Stream (Of a) m r -> Source a m r
 
      But see 'uncons', which is better fitted to these @unfoldM@s
 -}
@@ -1434,21 +1431,21 @@ next = loop where
 
 notElem :: (Monad m, Eq a) => a -> Stream (Of a) m r -> m (Of Bool r)
 notElem a' = loop True where
-  loop False str = liftM (False :>) (effects str)
+  loop False str = fmap (False :>) (effects str)
   loop True str = case str of
     Return r -> return (True:> r)
     Effect m -> m >>= loop True
     Step (a:> rest) ->
       if a == a'
-        then liftM (False :>) (effects rest)
+        then fmap (False :>) (effects rest)
         else loop True rest
 {-#INLINABLE notElem #-}
 
 notElem_ :: (Monad m, Eq a) => a -> Stream (Of a) m r -> m Bool
 notElem_ a' = loop True where
-  loop False str = return False
+  loop False _ = return False
   loop True str = case str of
-    Return r -> return True
+    Return _ -> return True
     Effect m -> m >>= loop True
     Step (a:> rest) ->
       if a == a'
@@ -1465,7 +1462,7 @@ partition :: Monad m => (a -> Bool) -> Stream (Of a) m r -> Stream (Of a) (Strea
 partition thus = loop where
    loop str = case str of
      Return r -> Return r
-     Effect m -> Effect (liftM loop (lift m))
+     Effect m -> Effect (fmap loop (lift m))
      Step (a :> rest) -> if thus a
        then Step (a :> loop rest)
        else Effect $ do
@@ -1497,7 +1494,7 @@ partitionEithers :: Monad m => Stream (Of (Either a b)) m r -> Stream (Of a) (St
 partitionEithers =  loop where
    loop str = case str of
      Return r -> Return r
-     Effect m -> Effect (liftM loop (lift m))
+     Effect m -> Effect (fmap loop (lift m))
      Step (Left a :> rest) -> Step (a :> loop rest)
      Step (Right b :> rest) -> Effect $ do
        yield b
@@ -1577,7 +1574,7 @@ repeatM ma = loop where
 
 -- | Repeat an element several times.
 replicate :: Monad m => Int -> a -> Stream (Of a) m ()
-replicate n a | n <= 0 = return ()
+replicate n _ | n <= 0 = return ()
 replicate n a = loop n where
   loop 0 = Return ()
   loop m = Effect (return (Step (a :> loop (m-1))))
@@ -1591,12 +1588,12 @@ replicate n a = loop n where
 
 -}
 replicateM :: Monad m => Int -> m a -> Stream (Of a) m ()
-replicateM n ma | n <= 0 = return ()
+replicateM n _ | n <= 0 = return ()
 replicateM n ma = loop n where
   loop 0 = Return ()
-  loop n = Effect $ do
+  loop m = Effect $ do
     a <- ma
-    return (Step (a :> loop (n-1)))
+    return (Step (a :> loop (m-1)))
 {-# INLINABLE replicateM #-}
 
 {-| Read an @IORef (Maybe a)@ or a similar device until it reads @Nothing@.
@@ -1640,7 +1637,7 @@ scan step begin done str = Step (done begin :> loop begin str)
   loop !acc stream = do
     case stream of
       Return r -> Return r
-      Effect m -> Effect (liftM (loop acc) m)
+      Effect m -> Effect (fmap (loop acc) m)
       Step (a :> rest) -> 
         let !acc' = step acc a 
         in Step (done acc' :> loop acc' rest)
@@ -1701,7 +1698,7 @@ scanned step begin done = loop Nothing' begin
     loop !m !x stream = do
       case stream of
         Return r -> return r
-        Effect mn  -> Effect $ liftM (loop m x) mn
+        Effect mn  -> Effect $ fmap (loop m x) mn
         Step (a :> rest) -> do
           case m of
             Nothing' -> do
@@ -1757,7 +1754,7 @@ sequence :: Monad m => Stream (Of (m a)) m r -> Stream (Of a) m r
 sequence = loop where
   loop stream = case stream of
     Return r          -> Return r
-    Effect m           -> Effect $ liftM loop m
+    Effect m           -> Effect $ fmap loop m
     Step (ma :> rest) -> Effect $ do
       a <- ma
       return (Step (a :> loop rest))
@@ -1810,11 +1807,11 @@ sum = fold (+) 0 id
 -- | Stream elements until one fails the condition, return the rest.
 span :: Monad m => (a -> Bool) -> Stream (Of a) m r
       -> Stream (Of a) m (Stream (Of a) m r)
-span pred = loop where
+span thePred = loop where
   loop str = case str of
     Return r         -> Return (Return r)
-    Effect m          -> Effect $ liftM loop m
-    Step (a :> rest) -> if pred a
+    Effect m          -> Effect $ fmap loop m
+    Step (a :> rest) -> if thePred a
       then Step (a :> loop rest)
       else Return (Step (a :> rest))
 {-# INLINABLE span #-}
@@ -1834,7 +1831,7 @@ split :: (Eq a, Monad m) =>
 split t  = loop  where
   loop stream = case stream of
     Return r -> Return r
-    Effect m -> Effect (liftM loop m)
+    Effect m -> Effect (fmap loop m)
     Step (a :> rest) ->
          if a /= t
             then Step (fmap loop (yield a >> break (== t) rest))
@@ -1842,9 +1839,9 @@ split t  = loop  where
 {-#INLINABLE split #-}
 
 {-| Split a succession of layers after some number, returning a streaming or
---   effectful pair. This function is the same as the 'splitsAt' exported by the
---   @Streaming@ module, but since this module is imported qualified, it can
---   usurp a Prelude name. It specializes to:
+    effectful pair. This function is the same as the 'splitsAt' exported by the
+    @Streaming@ module, but since this module is imported qualified, it can
+    usurp a Prelude name. It specializes to:
 
 >  splitAt :: (Monad m, Functor f) => Int -> Stream (Of a) m r -> Stream (Of a) m (Stream (Of a) m r)
 
@@ -1869,7 +1866,7 @@ subst :: (Monad m, Functor f) =>  (a -> f x) -> Stream (Of a) m r -> Stream f m 
 subst f s = loop s where
   loop str = case str of
     Return r         -> Return r
-    Effect m         -> Effect (liftM loop m)
+    Effect m         -> Effect (fmap loop m)
     Step (a :> rest) -> Step (loop rest <$ f a)
 {-#INLINABLE subst #-}
 -- ---------------
@@ -1894,13 +1891,14 @@ import Streaming.Prelude (each, next, yield)
 -}
 
 take :: (Monad m, Functor f) => Int -> Stream f m r -> Stream f m ()
-take n0 str | n0 <= 0 = return ()
+take n0 _ | n0 <= 0 = return ()
 take n0 str = loop n0 str where
-  loop 0 p = return ()
+  loop 0 _ = return ()
   loop n p =
-    case p of Step fas -> Step (fmap (loop (n-1)) fas)
-              Effect m -> Effect (liftM (loop n) m)
-              Return r -> Return ()
+    case p of
+      Step fas -> Step (fmap (loop (n-1)) fas)
+      Effect m -> Effect (fmap (loop n) m)
+      Return _ -> Return ()
 {-# INLINABLE take #-}
 
 -- ---------------
@@ -1923,22 +1921,22 @@ take n0 str = loop n0 str where
 
 -}
 takeWhile :: Monad m => (a -> Bool) -> Stream (Of a) m r -> Stream (Of a) m ()
-takeWhile pred = loop where
+takeWhile thePred = loop where
   loop str = case str of
-    Step (a :> as) -> when (pred a) (Step (a :> loop as))
-    Effect m              -> Effect (liftM loop m)
-    Return r              -> Return ()
+    Step (a :> as) -> when (thePred a) (Step (a :> loop as))
+    Effect m -> Effect (fmap loop m)
+    Return _ -> Return ()
 {-# INLINE takeWhile #-}
 
 {-| Like 'takeWhile', but takes a monadic predicate. -}
 takeWhileM :: Monad m => (a -> m Bool) -> Stream (Of a) m r -> Stream (Of a) m ()
-takeWhileM pred = loop where
+takeWhileM thePred = loop where
   loop str = case str of
     Step (a :> as) -> do
-      b <- lift (pred a)
+      b <- lift (thePred a)
       when b (Step (a :> loop as))
-    Effect m -> Effect (liftM loop m)
-    Return r -> Return ()
+    Effect m -> Effect (fmap loop m)
+    Return _ -> Return ()
 {-# INLINE takeWhileM #-}
 
 
@@ -2067,7 +2065,7 @@ with :: (Monad m, Functor f) => Stream (Of a) m r -> (a -> f x) -> Stream f m r
 with s f = loop s where
   loop str = case str of
     Return r         -> Return r
-    Effect m         -> Effect (liftM loop m)
+    Effect m         -> Effect (fmap loop m)
     Step (a :> rest) -> Step (loop rest <$ f a)
 {-#INLINABLE with #-}
 
@@ -2117,10 +2115,10 @@ zipWith f = loop
   where
     loop str0 str1 = case str0 of
       Return r          -> Return r
-      Effect m           -> Effect $ liftM (\str -> loop str str1) m
+      Effect m           -> Effect $ fmap (\str -> loop str str1) m
       Step (a :> rest0) -> case str1 of
         Return r          -> Return r
-        Effect m           -> Effect $ liftM (loop str0) m
+        Effect m           -> Effect $ fmap (loop str0) m
         Step (b :> rest1) -> Step (f a b :>loop rest0 rest1)
 {-# INLINABLE zipWith #-}
 
@@ -2587,7 +2585,7 @@ copy
 copy = Effect . return . loop where
   loop str = case str of
     Return r         -> Return r
-    Effect m         -> Effect (liftM loop (lift m))
+    Effect m         -> Effect (fmap loop (lift m))
     Step (a :> rest) -> Effect (Step (a :> Return (Step (a :> loop rest))))
 {-#INLINABLE copy#-}
 
@@ -2654,7 +2652,7 @@ unzip :: Monad m =>  Stream (Of (a,b)) m r -> Stream (Of a) (Stream (Of b) m) r
 unzip = loop where
  loop str = case str of
    Return r -> Return r
-   Effect m -> Effect (liftM loop (lift m))
+   Effect m -> Effect (fmap loop (lift m))
    Step ((a,b):> rest) -> Step (a :> Effect (Step (b :> Return (loop rest))))
 {-#INLINABLE unzip #-}
 
@@ -2673,7 +2671,7 @@ catMaybes :: Monad m => Stream (Of (Maybe a)) m r -> Stream (Of a) m r
 catMaybes = loop where
   loop stream = case stream of
     Return r -> Return r
-    Effect m -> Effect (liftM loop m)
+    Effect m -> Effect (fmap loop m)
     Step (ma :> snext) -> case ma of
       Nothing -> loop snext
       Just a -> Step (a :> loop snext)
@@ -2688,7 +2686,7 @@ mapMaybe :: Monad m => (a -> Maybe b) -> Stream (Of a) m r -> Stream (Of b) m r
 mapMaybe phi = loop where
   loop stream = case stream of
     Return r -> Return r
-    Effect m -> Effect (liftM loop m)
+    Effect m -> Effect (fmap loop m)
     Step (a :> snext) -> case phi a of
       Nothing -> loop snext
       Just b -> Step (b :> loop snext)
@@ -2722,9 +2720,22 @@ slidingWindow n = setup (max 1 n :: Int) mempty
     setup 0 !sequ str = do
        yield sequ 
        window (Seq.drop 1 sequ) str 
-    setup n sequ str = do 
+    setup m sequ str = do 
       e <- lift $ next str 
       case e of 
         Left r ->  yield sequ >> return r
-        Right (x,rest) -> setup (n-1) (sequ Seq.|> x) rest
+        Right (x,rest) -> setup (m-1) (sequ Seq.|> x) rest
 {-#INLINABLE slidingWindow #-}
+
+-- | Map monadically over a stream, producing a new stream
+--   only containing the 'Just' values.
+mapMaybeM :: Monad m => (a -> m (Maybe b)) -> Stream (Of a) m r -> Stream (Of b) m r
+mapMaybeM phi = loop where
+  loop stream = case stream of
+    Return r -> Return r
+    Effect m -> Effect (fmap loop m)
+    Step (a :> snext) -> Effect $ do
+      flip fmap (phi a) $ \x -> case x of
+        Nothing -> loop snext
+        Just b -> Step (b :> loop snext)
+{-#INLINABLE mapMaybeM #-}